refactor(csharp): centralize type decisions on CSharpType, enforce one-way module deps
